+++ Wrappers/Python/test/test_CIL_CVXPy.py
@@ -21,6 +21,7 @@ import unittest
 from utils import initialise_tests
 from cil.optimisation.functions import L2NormSquared
 from cil.optimisation.functions import TotalVariation
+from cil.optimisation.functions import TotalGeneralisedVariation
 from cil.utilities import dataexample
 import numpy as np
 import scipy.sparse as sp
@@ -218,42 +219,53 @@ class Test_CIL_vs_CVXPy(unittest.TestCase):
             # compare objectives
             f = 0.5*L2NormSquared(b=self.data)
             cil_objective = f(tv_cil) + TV(tv_cil)*(3)
-            np.testing.assert_allclose(cil_objective, obj.value, atol=1e-3) 
+            np.testing.assert_allclose(cil_objective, obj.value, atol=1e-3)    
 
     @unittest.skipUnless(has_cvxpy, "CVXpy not installed")
-    def test_cil_vs_cvxpy_totalvariation_strongly_convex(self):  
+    def tgv_cvxpy_regulariser(self, u, w1, w2, alpha1, alpha0, boundaries = "Neumann"):
 
+        G1 = self.sparse_gradient_matrix(u.shape, direction = 'forward', order = 1, boundaries = boundaries)  
+        DX, DY = G1[1], G1[0]
+
+        G2 = self.sparse_gradient_matrix(u.shape, direction = 'backward', order = 1, boundaries = boundaries) 
+        divX, divY = G2[1], G2[0]
+
+        return alpha1 * cvxpy.sum(cvxpy.norm(cvxpy.vstack([DX @ cvxpy.vec(u) - cvxpy.vec(w1), DY @ cvxpy.vec(u) - cvxpy.vec(w2)]), 2, axis = 0)) + \
+            alpha0 * cvxpy.sum(cvxpy.norm(cvxpy.vstack([ divX @ cvxpy.vec(w1), divY @ cvxpy.vec(w2), \
+                                        0.5 * ( divX @ cvxpy.vec(w2) + divY @ cvxpy.vec(w1) ), \
+                                        0.5 * ( divX @ cvxpy.vec(w2) + divY @ cvxpy.vec(w1) ) ]), 2, axis = 0  ) )           
+
+    @unittest.skipUnless(has_cvxpy, "CVXpy not installed")                                    
+    def test_cil_vs_cvxpy_total_generalised_variation(self):
+        
         # solution
         u_cvx = cvxpy.Variable(self.data.shape)
+        w1_cvx = cvxpy.Variable(self.data.shape)
+        w2_cvx = cvxpy.Variable(self.data.shape)
 
-        # regularisation parameter
-        alpha = 0.1
-
-        # strongly convex constant
-        gamma = 0.05
+        # regularisation parameters
+        alpha1 = 0.1
+        alpha0 = 0.5
 
         # fidelity term
-        fidelity = 0.5 * cvxpy.sum_squares(u_cvx - self.data.array) 
-        regulariser = alpha * self.tv_cvxpy_regulariser(u_cvx) +  (gamma/2) * cvxpy.sum_squares(u_cvx)
+        fidelity = 0.5 * cvxpy.sum_squares(u_cvx - self.data.array)   
+        regulariser = self.tgv_cvxpy_regulariser(u_cvx, w1_cvx, w2_cvx, alpha1, alpha0)
 
         # objective
         obj =  cvxpy.Minimize( regulariser +  fidelity)
         prob = cvxpy.Problem(obj, constraints = [])
 
-        # Choose solver ( SCS, MOSEK(license needed) )
-        tv_cvxpy = prob.solve(verbose = True, solver = cvxpy.SCS)   
+        # Choose solver (SCS)
+        tvg_cvxpy = prob.solve(verbose = True, solver = cvxpy.SCS)   
 
-        # use TotalVariation from CIL (with Fast Gradient Projection algorithm)
-        TV = alpha * TotalVariation(max_iteration = 500, strong_convexity_constant = gamma)
-        tv_cil = TV.proximal(self.data, tau=1.0)                
-
+        TGV = TotalGeneralisedVariation(alpha1 = alpha1, alpha0 = alpha0, max_iteration=2000, verbose=0)
+        tgv_cil = TGV.proximal(self.data, tau = 1.0)
+        
         # compare solution
-        np.testing.assert_allclose(tv_cil.array, u_cvx.value, atol=1e-2)                           
+        np.testing.assert_allclose(tgv_cil.array, u_cvx.value, atol=1e-1) 
+        np.testing.assert_allclose(TGV.pdhg.objective[-1], obj.value, atol=1e-1)   
 
-        # compare objectives
-        f = 0.5*L2NormSquared(b=self.data)
-        cil_objective = f(tv_cil) + TV(tv_cil) 
-        np.testing.assert_allclose(cil_objective, obj.value, atol=1e-1)         
+